requirements.txt
fastapi[standard]>=0.114.2,<1.0.0
python-multipart>=0.0.7,<1.0.0
email-validator>=2.1.0.post1,<3.0.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
tenacity>=8.2.3,<9.0.0
pydantic>2.0
emails>=0.6,<1.0
jinja2>=3.1.4,<4.0.0
alembic>=1.12.1,<2.0.0
httpx>=0.25.1,<1.0.0
psycopg[binary]>=3.1.13,<4.0.0
sqlmodel>=0.0.21,<1.0.0
bcrypt==4.0.1
pydantic-settings>=2.2.1,<3.0.0
sentry-sdk[fastapi]>=1.40.6,<2.0.0
pyjwt>=2.8.0,<3.0.0
uvicorn[standard]>=0.30.0,<1.0.0
gunicorn>=21.2.0
